refactor(shared): dedupe outbound URL host validation
Share the parse-original and host/IP validation path between webhook and LLM base URL validation, and cover the LLM parser mismatch regression.
